% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacant_helpers.R
\name{acat_t}
\alias{acat_t}
\title{Calculate ACAT test statistic}
\usage{
acat_t(p, w)
}
\arguments{
\item{p}{Numeric vector of p-values (each in (0,1)).}

\item{w}{Numeric vector of positive weights of the same length as \code{p}.}
}
\value{
Numeric scalar: the ACAT test statistic.
}
\description{
Given a vector of p-values and corresponding weights, computes the aggregated
Cauchy test statistic.
}
