% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_set.R
\name{analyze_set}
\alias{analyze_set}
\title{Analyze score-based clusters with Firth regression}
\usage{
analyze_set(
  pre,
  clus,
  test = c("uni", "multi"),
  acat.weight = c("score", "equal")
)
}
\arguments{
\item{pre}{List. Pre-processing outputs (phenotype, genotypes, null model info).}

\item{clus}{List. Output from \code{cluster_score()}, containing assignments
and calculated scalar weights (\code{ac.weights}).}

\item{test}{Character. "uni" or "multi".}

\item{acat.weight}{Character. "score" or "equal".}
}
\value{
Named numeric vector with results.
}
\description{
Performs either sequential univariate tests or a joint multivariate test.
}
