% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_summary.R
\name{print.vacant_tier_summary}
\alias{print.vacant_tier_summary}
\title{Print method for VACANT tier summary}
\usage{
\method{print}{vacant_tier_summary}(x, ...)
}
\arguments{
\item{x}{A \code{vacant_tier_summary} object.}

\item{...}{Ignored.}
}
\description{
Print method for VACANT tier summary
}
