% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacant_helpers.R
\name{extract_sub}
\alias{extract_sub}
\title{Extract a single character from a genotype string}
\usage{
extract_sub(x, idx)
}
\arguments{
\item{x}{Character vector of genotype strings.}

\item{idx}{Integer. Position of allele to extract.}
}
\value{
Character scalar of the extracted allele.
}
\description{
Given a genotype string vector and an index, returns the allele at that position.
}
