% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vacant_cluster.R
\name{predict_vacant_cluster}
\alias{predict_vacant_cluster}
\title{Predict cluster assignment for new variants using the Pareto Staircase Model}
\usage{
predict_vacant_cluster(model, new.scores)
}
\arguments{
\item{model}{The prediction model object returned by \code{cluster_score()}.}

\item{new.scores}{A data frame or matrix of raw scores for new variants.}
}
\value{
An integer vector of cluster assignments.
}
\description{
Classifies new variants into risk clusters based on the "Layered Pareto Staircase"
model generated by \code{cluster_score}.
}
