% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacant.R
\name{vacant}
\alias{vacant}
\title{Run the VACANT Analysis (Returning Model Object)}
\usage{
vacant(
  geno.maf,
  score.vector,
  phenotype,
  covariates = NULL,
  test = c("uni", "multi"),
  acat.weight = c("score", "equal"),
  size.threshold = 10
)
}
\value{
A list containing analysis results ($results) and the prediction model ($model).
}
\description{
Run the VACANT Analysis (Returning Model Object)
}
